- move functions graph_from_file and graph_from_molfile_text from molfile_v3000_reader to the more generic module molfile_reader - MolfileParserException needs to be moved to a separate file due to circular dependency clashes
