+++ tucan/io/__init__.py
@@ -1,10 +1,11 @@
 """TUCAN IO package"""
 
-from tucan.io.molfile_v3000_reader import (
+from tucan.io.molfile_reader import (
     graph_from_file,
     graph_from_molfile_text,
-    MolfileParserException,
 )
+
+from tucan.io.exception import MolfileParserException
 
 from tucan.io.molfile_writer import graph_to_molfile
